moved some components
