+++ package.json
@@ -4,8 +4,6 @@
   "private": true,
   "homepage": "https://TrueMolni.github.io/your-pet-project-frontend/",
   "dependencies": {
-    "@emotion/react": "^11.11.0",
-    "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.11.16",
     "@reduxjs/toolkit": "^1.9.5",
     "@testing-library/jest-dom": "^5.16.3",
