+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,11 +1,11 @@
-import LoginForm from "components/AuthForm/LoginForm";
+import LoginForm from 'modules/AuthForm/LoginForm';
 const LoginPage = () => {
-    return (
-      <div>
-        {/* <p>login page (temp sign)</p>
+  return (
+    <div>
+      {/* <p>login page (temp sign)</p>
       <h1>Login</h1> */}
-        <LoginForm title={("Login")} />
-      </div>
-    );
-}
+      <LoginForm title={'Login'} />
+    </div>
+  );
+};
 export default LoginPage;
